% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{adjMats}
\alias{adjMats}
\title{Construct adjacency matrices for fixed and random effects}
\usage{
adjMats(self)
}
\arguments{
\item{self}{LARMExFit. A setRefClass object holding data and settings}
}
\value{
Updates $feM and $reM based on $arAbb and $exAbb of self
}
\description{
Construct adjacency matrices for fixed and random effects
}
