% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{doAbbreviate}
\alias{doAbbreviate}
\title{Abbreviate long names according to self$nAbb}
\usage{
doAbbreviate(self)
}
\arguments{
\item{self}{LARMExFit. A setRefClass object holding data and settings}
}
\value{
Updates $arAbb and $exAbb based on $arList, $exList and $nAbb.
If $nAbb is zero or greater than the lengths of all names, the original
names are retained.
}
\description{
Abbreviate long names according to self$nAbb
}
\examples{
obj1 <-LARMExFit$new()
obj1$arList <- c("mood1", "mood2")
doAbbreviate(obj1)
}
