% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{fitLmer1}
\alias{fitLmer1}
\title{Fit a LARMEx model to the data of a single subject}
\usage{
fitLmer1(self, sj, toDir)
}
\arguments{
\item{self}{LARMExFit. n R6Class object holding data and settings}

\item{sj}{ID for a single subject}

\item{toDir}{Path to save results}
}
\description{
Fit a LARMEx model to the data of a single subject
}
\examples{
obj1 <- LARMExFit$new()
fpath <- system.file("extdata", "simMood.csv", package="larmexShiny")
obj1$rawData <- read.csv(fpath)
obj1$sjID <- "sjID"
obj1$nDay <- "nDay"
obj1$nBeep <- "nBeep"
obj1$arList <- c("M1", "M2")
obj1$exList <- "E"
obj1$savePath <- normalizePath("~")
selCols <- c(obj1$nDay, obj1$nBeep, obj1$arList, obj1$exList)
toDir = file.path(obj1$savePath, "fitRes")
dir.create(toDir, recursive=TRUE, showWarnings=FALSE)
sj <- 1001
sjData <- obj1$rawData[obj1$rawData[obj1$sjID]==sj, selCols]
df2Fit <- prepData2Fit(obj1, sjData)
setFormulaDf(obj1, df2Fit)
fitLmer1(obj1, sj, toDir)
}
