% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{LARMExFit}
\alias{LARMExFit}
\title{Collection of data and settings}
\description{
An R6Class is used instead of a named list
because we needed a data structure to be passed to functions by reference
instead of value. It is not implemented in a pure object oriented form for
simplicity. However, it is kept similar by naming this object "self"
and passing it as an argument to independent functions.
}
\examples{
obj1 <-LARMExFit$new()
obj1$arList <- c("mood1", "mood2")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rawData}}{data, single or multiple subjects, long dataframe}

\item{\code{sjID}}{variable name, subject ID, character}

\item{\code{nBeep}}{variable name, number of beeps, character}

\item{\code{nDay}}{variable name, number of days, character}

\item{\code{arList}}{list of interacting mood nodes, character}

\item{\code{exList}}{list of external factors acting on moods, character}

\item{\code{nAbb}}{abbreviate long names to this length, ignore if 0}

\item{\code{arAbb}}{arList after abbreviation}

\item{\code{exAbb}}{exlist after abbreviation}

\item{\code{fitID}}{list of subject of IDs to be fitted}

\item{\code{frm}}{"lmer" formula to be fit}

\item{\code{fitRes}}{result of fitting}

\item{\code{feM}}{adjacency matrix of fixed effects}

\item{\code{reM}}{adjacency matrix of random effects}

\item{\code{savePath}}{path to save results, default to HOME}

\item{\code{doCenter}}{center mood and external factor values per subject}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LARMExFit-clone}{\code{LARMExFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LARMExFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LARMExFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LARMExFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
