% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{prepData2Fit}
\alias{prepData2Fit}
\title{Transform raw data to a format suitable for fitting by LARMEx}
\usage{
prepData2Fit(self, sjData)
}
\arguments{
\item{self}{LARMExFit. A setRefClass object holding data and settings}

\item{sjData}{Raw data for single subject}
}
\value{
data.frame. Transformed data
}
\description{
Transform raw data to a format suitable for fitting by LARMEx
}
\examples{
obj1 <- LARMExFit$new()
fpath <- system.file("extdata", "simMood.csv", package="larmexShiny")
obj1$rawData <- read.csv(fpath)
obj1$nDay <- "nDay"
obj1$nBeep <- "nBeep"
obj1$arList <- c("M1", "M2")
obj1$exList <- "E"
selCols <- c(obj1$nDay, obj1$nBeep, obj1$arList, obj1$exList)
sj <- 1001
sjData <- obj1$rawData[obj1$rawData[obj1$sjID]==sj, selCols]
df2Fit <- prepData2Fit(obj1,sjData)
}
