% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{propperLevels}
\alias{propperLevels}
\title{Check if data has a proper multilevel form for the Shiny app}
\usage{
propperLevels(self, nD = 2, nB = 2)
}
\arguments{
\item{self}{LARMExFit. An R6Class object holding data and settings}

\item{nD}{integer. Minimum number of days per subject}

\item{nB}{integer. Minimum number of beeps per day}
}
\description{
Check if data has a proper multilevel form for the Shiny app
}
