% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{setFormulaDf}
\alias{setFormulaDf}
\title{Generate mixed-effects formula from a transformed data frame
see help('prepData2Fit')}
\usage{
setFormulaDf(self, data2Fit)
}
\arguments{
\item{self}{LARMExFit. A setRefClass object holding data and settings}

\item{data2Fit}{the formula is constructed from its columns of prepared data}
}
\value{
self$frm is updated
}
\description{
Generate mixed-effects formula from a transformed data frame
see help('prepData2Fit')
}
\examples{
obj1 <- LARMExFit$new()
fpath <- system.file("extdata", "simMood.csv", package="larmexShiny")
obj1$rawData <- read.csv(fpath)
obj1$nDay <- "nDay"
obj1$nBeep <- "nBeep"
obj1$arList <- c("M1", "M2")
obj1$exList <- "E"
selCols <- c(obj1$nDay, obj1$nBeep, obj1$arList, obj1$exList)
sj <- 1001
sjData <- obj1$rawData[obj1$rawData[obj1$sjID]==sj, selCols]
df2Fit <- prepData2Fit(obj1, sjData)
setFormulaDf(obj1, df2Fit)
}
