% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARMExFit.R
\name{setFormulaM}
\alias{setFormulaM}
\title{Generate mixed-effects formula from adjacency matrices of fixed and
random effects given that "self" has at least two variables in $arList}
\usage{
setFormulaM(self)
}
\arguments{
\item{self}{LARMExFit. A setRefClass object holding data and settings}
}
\value{
self$frm is updated
}
\description{
Generate mixed-effects formula from adjacency matrices of fixed and
random effects given that "self" has at least two variables in $arList
}
\examples{
obj1 <-LARMExFit$new()
obj1$nDay <- "nDay"
obj1$arList <- c("M1", "M2")
obj1$exList <- "E"
adjMats(obj1)
setFormulaM(obj1)
}
